refactor(THU-148): extract LocationSearchCombobox into shared component
- deduplicate location search popover from onboarding and preferences
- move useLocationSearch hook usage into the new component
- add autoOpen prop for onboarding auto-focus behavior
